+++ frontend/src/components/general/AuthForms.jsx
@@ -1,22 +1,15 @@
 import LoginForm from "./LoginForm";
 import SignupModal from "./SignupModal";
 
-function AuthForms(props) {
-  // eslint-disable-next-line react/prop-types
-  const {setUserInfo} = props;
-
+function AuthForms() {
   return (
     <>
       {/* ログインフォーム */}
       <div>
         <p className="text-lg text-center font-semibold mb-3">ログイン</p>
-        <LoginForm
-          setUserInfo={setUserInfo}
-        />
+        <LoginForm />
         <div className="divider text-xs text-gray-500 font-medium my-3">OR</div>
-        <SignupModal
-          setUserInfo={setUserInfo}
-        />
+        <SignupModal />
       </div>
     </>
   )
